Extract ranking table component
